% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqmap.R
\name{freqmap}
\alias{freqmap}
\title{Get Frequency Count of kmers}
\arguments{
\item{dnastr}{- dna string}

\item{k}{- length of the k-mer}
}
\value{
A hash map of k-mers and their counts
}
\description{
Get Frequency Count of kmers
}
