% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontent.R
\name{GCcontent}
\alias{GCcontent}
\title{The GC Content}
\arguments{
\item{dnastr}{- the dna string}
}
\value{
A fraction representing the (G+C) portion of the dnastr
}
\description{
The GC Content
}
\examples{


}
