% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{hamming}
\alias{hamming}
\title{Hamming Distance}
\arguments{
\item{pattern1}{- pattern1 to compare with pattern2}

\item{pattern2}{- pattern2 of equal length with pattern1}
}
\value{
a number showing the number of differences between the patterns
}
\description{
Hamming Distance
}
\examples{

}
