% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearmatches.R
\name{nearmatches}
\alias{nearmatches}
\title{Near Pattern Match Count}
\usage{
nearmatches(pattern, dnastr, d)
}
\arguments{
\item{pattern}{- pattern to search for}

\item{dnastr}{-   dna string to search the pattern in}

\item{d}{- Number of differences away the closest match can be}
}
\value{
a number showing the number of near matches for the pattern with d deviations
}
\description{
Near Pattern Match Count
}
\examples{

}
