% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterncount.R
\name{patterncount}
\alias{patterncount}
\title{Count of the Given Pattern}
\arguments{
\item{dnastr}{- a string containing the dna sequence to be analyzed}

\item{pattern}{- pattern to look for}
}
\value{
a number showing the count of the occurrences of the pattern
}
\description{
Count of the Given Pattern
}
\examples{



}
