% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternpositions.R
\name{patternpositions}
\alias{patternpositions}
\title{Positions of the Pattern in dna String}
\arguments{
\item{dnastr}{- a string containing the dna sequence to be analyzed}

\item{pattern}{- pattern to look for}
}
\value{
a vector of the positions of the occurrences of the pattern in the dnastr
}
\description{
Positions of the Pattern in dna String
}
\examples{

}
