% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsimpledna.R
\name{testsimpledna}
\alias{testsimpledna}
\title{test-simpledna}
\usage{
testsimpledna()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
This is a test script file for testing the functions of the simpledna R package.
}
\details{
It uses a test dataset den-1.fasta.txt which is the complete genome of Dengue virus 1
in fasta format
}
